+++ CHANGELOG.md
@@ -27,6 +27,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - fixed an issue where Oxidized could not pull config from XOS-devices operating in stacked mode (@DarkCatapulter)
 - fixed an issue where Oxidized could not pull config from XOS-devices that have not saved their configuration (@DarkCatapulter)
 - improved scrubbing of show chassis in ironware model (@michaelpsomiadis)
+- fixed an issue where Oxidized was unable to match post login prompt RegEx in new DmOS firmware (@Zuquim)
+- fixed an issue where Oxidized was unable to properly paginate DataCom devices' terminal output (@Zuquim)
 
 ## [0.28.0 - 2020-05-18]